requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=2.0.0
gymnasium>=0.28.0
nle>=0.9.0

# Hypergraph and embeddings
networkx>=2.8
scikit-learn>=1.2.0
requests>=2.28.0

# LLM
openai>=1.0.0
anthropic>=0.7.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
tensorboard>=2.13.0
matplotlib>=3.5.0
seaborn>=0.11.0
